Leaderboard Entry - Modifying LevelData XML
UpdateLevelData class that will handle loading the current leaderboard,
check a new time against and inserting it if good enough and saving this
new leaderboard.

Currently not tested or implented - Only logic plus comments

Also editted the folder url in MenuLoadLevelsFromXML.cs to be able to
find the dataPath of the game and find levels from there rather than
manually adjusting the file path in code.