Level Files Now Checked In Menu
Level files are checked to be valid level files. This is defined as a
file with a Stats, PlayerStart, Goal, Platforms, Targets and Towers node
inside a LevelData node. If any files are returned false for validity,
then an error message will spring up in the menu to say how many levels
were invalid.

Also edited the current XML level files to valid to the new checks.